SRF -> EXE Merge
In order to avoid issues in Linux with needing the SRF in the same directory as the executable, I've made SCI able to get the SRF info from itself *if* SRF is actually merged. this way the SCI_Run executable may just be in /usr/bin like all executables and all assets belonging to SCI itself are all with it. Game projects releases in Linux may then just be the main JCR6 file which just requires SCI as a dependency. Now creating packages for package managers never had much PROPER and EASY TO UNDERSTAND documentations (since in Linux things must always be needlessly complicated as apparently that's "efficicient"), but it is still a field I want to explore someday. When it comes to .AppImages in which I also have a great interest, this won't really matter... or does it?